formaDePago

+++ src/main/java/com/juandevs/prue11/service/FormaDePagoServiceImpl.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Service;
 
 import com.juandevs.prue11.entity.FormasDePago;
 import com.juandevs.prue11.repository.FormaDePagoRepository;
+import com.juandevs.prue11.request.Response;
 import com.juandevs.prue11.service.interfaces.IFormaDePagoService;
 
 
@@ -17,23 +18,73 @@ public class FormaDePagoServiceImpl implements IFormaDePagoService{
     FormaDePagoRepository formaDePagoRepository;
 
     @Override
-    public Iterable<FormasDePago> findAll() {
-        return formaDePagoRepository.findAll();
+    public Response<Iterable<FormasDePago>> findAll() {
+        try {
+            return new Response<Iterable<FormasDePago>>("Consulta realizada con exito", true, formaDePagoRepository.findAll());
+        } catch (Exception e) {
+            return new Response<Iterable<FormasDePago>>(e.getMessage(), false, null);
+        }
     }
 
     @Override
-    public Optional<FormasDePago> findById(int id) {
-        return formaDePagoRepository.findById(id);
+    public Response<Optional<FormasDePago>> findById(int id) {
+        try {
+            Optional<FormasDePago> producto = formaDePagoRepository.findById(id);
+
+            if(!producto.isPresent()) return new Response<Optional<FormasDePago>>("Producto no encontrado",false,producto);
+        
+            return new Response<Optional<FormasDePago>>("FormasDePago encontrado",false,producto);
+        } catch (Exception e) {
+            return new Response<Optional<FormasDePago>>(e.getMessage(),false,null);
+        }
     }
 
     @Override
-    public FormasDePago save(FormasDePago formaDePago) {
-        return formaDePagoRepository.save(formaDePago);
+    public Response<FormasDePago> save(FormasDePago formasDePago) {
+        try {
+            if(formasDePago == null) {
+                return new Response<FormasDePago>("Los datos son nulos", false, formasDePago);
+            }
+    
+            if(formaDePagoRepository.findById(formasDePago.getId()).isPresent()) {
+                return new Response<FormasDePago>("La forma de pago ya se encuentra registrado", false, formasDePago);
+            }
+    
+            return new Response<FormasDePago>("Forma de pago guardado", true, formaDePagoRepository.save(formasDePago));
+        } catch (Exception e) {
+            return new Response<FormasDePago>(e.getMessage(), false, formasDePago);
+        }
     }
 
     @Override
-    public void deleteById(int id) {
-        formaDePagoRepository.deleteById(id);
+    public Response<FormasDePago> update(FormasDePago formaDepago) {
+        try {
+            if(formaDepago == null) {
+                return new Response<FormasDePago>("Los datos son nulos", false, formaDepago);
+            }
+    
+            if(!formaDePagoRepository.findById(formaDepago.getId()).isPresent()) {
+                return new Response<FormasDePago>("El producto no se encuentra registrado", false, formaDepago);
+            }
+    
+            return new Response<FormasDePago>("Forma de pago actualizada", true, formaDePagoRepository.save(formaDepago));
+        } catch (Exception e) {
+            return new Response<FormasDePago>(e.getMessage(), true, formaDepago);
+        }
+    }
+
+    @Override
+    public Response<Optional<FormasDePago>> deleteById(int id) {
+        try {
+            Optional<FormasDePago> formaDePago = formaDePagoRepository.findById(id);
+
+            if(!formaDePago.isPresent()) return new Response<Optional<FormasDePago>>("Forma de pago no encontrada",false,formaDePago);
+            
+            formaDePagoRepository.deleteById(id);
+            return new Response<Optional<FormasDePago>>("Formas de pago eliminado",false,formaDePago);
+        } catch (Exception e) {
+            return new Response<Optional<FormasDePago>>(e.getMessage(),false,null);
+        }
     }
     
 }
